requirements.txt
openai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
